requirements.txt
discord.py
asyncio
pytz
